Struct bloomfilter::Bloom

pub struct Bloom<T: ?Sized> { /* private fields */ }

Bloom filter structure

Implementations

impl<T: ?Sized> Bloom<T>

pub fn new_with_seed( bitmap_size: usize, items_count: usize, seed: &[u8; 32] ) -> Self

Create a new bloom filter structure. bitmap_size is the size in bytes (not bits) that will be allocated in memory items_count is an estimation of the maximum number of items to store. seed is a random value used to generate the hash functions.

pub fn new(bitmap_size: usize, items_count: usize) -> Self

Create a new bloom filter structure. bitmap_size is the size in bytes (not bits) that will be allocated in memory items_count is an estimation of the maximum number of items to store.

pub fn new_for_fp_rate(items_count: usize, fp_p: f64) -> Self

Create a new bloom filter structure. items_count is an estimation of the maximum number of items to store. fp_p is the wanted rate of false positives, in ]0.0, 1.0[

pub fn new_for_fp_rate_with_seed( items_count: usize, fp_p: f64, seed: &[u8; 32] ) -> Self

Create a new bloom filter structure. items_count is an estimation of the maximum number of items to store. fp_p is the wanted rate of false positives, in ]0.0, 1.0[

pub fn from_bit_vec( bit_vec: BitVec, bitmap_bits: u64, k_num: u32, sip_keys: [(u64, u64); 2] ) -> Self

Create a bloom filter structure from a previous state given as a ByteVec structure. The state is assumed to be retrieved from an existing bloom filter.

pub fn from_existing( bytes: &[u8], bitmap_bits: u64, k_num: u32, sip_keys: [(u64, u64); 2] ) -> Self

Create a bloom filter structure with an existing state given as a byte array. The state is assumed to be retrieved from an existing bloom filter.

pub fn compute_bitmap_size(items_count: usize, fp_p: f64) -> usize

Compute a recommended bitmap size for items_count items and a fp_p rate of false positives. fp_p obviously has to be within the ]0.0, 1.0[ range.

pub fn set(&mut self, item: &T)
where T: Hash,

Record the presence of an item.

pub fn check(&self, item: &T) -> bool
where T: Hash,

Check if an item is present in the set. There can be false positives, but no false negatives.

pub fn check_and_set(&mut self, item: &T) -> bool
where T: Hash,

Record the presence of an item in the set, and return the previous state of this item.

pub fn bitmap(&self) -> Vec<u8>

Return the bitmap as a vector of bytes

pub fn bit_vec(&self) -> &BitVec

Return the bitmap as a “BitVec” structure

pub fn number_of_bits(&self) -> u64

Return the number of bits in the filter

pub fn number_of_hash_functions(&self) -> u32

Return the number of hash functions used for check and set

pub fn sip_keys(&self) -> [(u64, u64); 2]

Return the keys used by the sip hasher

pub fn clear(&mut self)

Clear all of the bits in the filter, removing all keys from the set

pub fn fill(&mut self)

Set all of the bits in the filter, making it appear like every key is in the set

pub fn is_empty(&self) -> bool

Test if there are no elements in the set